patch from Stepan Mishura to ensure that we test the security code w/in the rest of the harmony code, using the Harmony VM - From HARMONY-58
Note that this isn't exact.  I removed the config of the boot classpath for 
the crypto and xnet becuase it's incomplete - it was relying on the security.jar
in JAVA_HOME/bin/lib/boot

So now, all code is from bin/lib/boot, and not local in modules/security2.  I'll
post a message to the mail list to suggest we modify this to use local.

What this means for now is that you must create the jars and have them
land in deploy/




git-svn-id: https://svn.apache.org/repos/asf/incubator/harmony/enhanced/classlib/trunk@376127 13f79535-47bb-0310-9956-ffa450edef68