fix to the problem reported in HARMONY-638
The problem was that fork() was failing due to 
what I can only assume was some kind of ulimit.
By running eclipse at the same time w/ a big 
heap, I can get this problem to show itself
on demand.

Problem was that execProgram() didn't check if
the fork() failed until after doing a blocking
read on a pipe to get a message from the child.
If the child was never created, it hung forever.

Now execProgram() returns 0 on success, and a
meaningful error code on failure, so that a 
nice Exception can be thrown.  The error codes
have to be split accross the platform impls of
this function.  I assume if people don't
reject this approach, we just put the error
codes into a common header.

Windows checkin will follow from other machine
after 1 svn update and smoke test.

We will need to find a regression test for
this and other stress-related maladies.



git-svn-id: https://svn.apache.org/repos/asf/incubator/harmony/enhanced/classlib/trunk@416412 13f79535-47bb-0310-9956-ffa450edef68