I occasionally run make in one of the native code sub-directories.  To do so correctly you need to define all of the variables that the make macro defines.  There are quite a lot and the list is growing.  So I've added a make.xml wrapper that gets copies in to the HDK so that you can do, for example:
  cd modules/luni/src/main/native/luni/unix
  HY_HDK=../../../../../../../deploy
  ant -f $HY_HDK/build/ant/make.xml

and it will do the equivalent of running make with all the correct flags.
The wrapper has the obvious rebuild, build and clean targets.
Be aware that you still need to specify the same -Dhy.no.thr=true, etc
flags to ant when invoking the wrapper that you normally would use -
or you can just add them to ~//.harmony-classlib.properties - because
these are not recorded in the HDK.


git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/classlib/trunk@667003 13f79535-47bb-0310-9956-ffa450edef68