Combine the strLen calculation and finding loops to optimise for the highly likely case that both will happen.
Fix off-by-one error that occurs because the strLen calculation and
the separator concatenation logic were not consistent.  This occurred
if you did something like:

  LD_LIBRARY_PATH=$PWD/deploy/jdk/jre/bin/default deploy/jdk/jre/bin/java ...

where the first newPathToAdd is missing but the second is found
leaving strLen with the wrong number of separators.  (It occurred for
me because the first newPathToAdd requires a trailing slash while the
second does not and I provided both without slashes.  We should fix
this too.)


git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/classlib/trunk@820201 13f79535-47bb-0310-9956-ffa450edef68