A bunch of new tests for OutputStreams.
This uses composition to create many tests (86 tests in this case) with few lines of code. This is most useful for things like OutputStream and Collection where there are multiple implementations that satisfy the same core behaviour. I intend to eventually create a similar set of testers for InputStreams, Readers and Writers. The main advantage of this style of test is that it becomes easy to check that all implementations conform to a specified behaviour.

The motivation for this specific set is a bug in FilteredStream.flush(), where we are ignoring an exception thrown by an underlying stream. These tests fail! There should be 7 new failures. I intend to commit a change that has the fix tomorrow.

git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/classlib/trunk@834317 13f79535-47bb-0310-9956-ffa450edef68